Update p_extnodes.c
p_extnodes.c: In function 'P_CheckMapFormat':
p_extnodes.c:46:37: warning: comparison of integer expressions of different signedness: 'int' and 'unsigned int' [-Wsign-compare]
   46 |     if ((b = lumpnum+ML_BLOCKMAP+1) < numlumps &&
      |                                     ^
p_extnodes.c:55:34: warning: comparison of integer expressions of different signedness: 'int' and 'unsigned int' [-Wsign-compare]
   55 |     if (!((b = lumpnum+ML_NODES) < numlumps &&
      |                                  ^
p_extnodes.c: In function 'P_LoadNodes_ZDBSP':
p_extnodes.c:369:29: warning: comparison of integer expressions of different signedness: 'unsigned int' and 'int' [-Wsign-compare]
  369 |     for (i = currSeg = 0; i < numsubsectors; i++)
      |                             ^
p_extnodes.c:394:19: warning: comparison of integer expressions of different signedness: 'unsigned int' and 'int' [-Wsign-compare]
  394 |     for (i = 0; i < numsegs; i++)
      |                   ^
p_extnodes.c:463:19: warning: comparison of integer expressions of different signedness: 'unsigned int' and 'int' [-Wsign-compare]
  463 |     for (i = 0; i < numnodes; i++)
      |                   ^

+++ src/doom/p_extnodes.c
@@ -41,7 +41,7 @@ mapformat_t P_CheckMapFormat (int lumpnum)
 {
     mapformat_t format = 0;
     byte *nodes = NULL;
-    int b;
+    unsigned int b;
 
     if ((b = lumpnum+ML_BLOCKMAP+1) < numlumps &&
         !strncasecmp(lumpinfo[b]->name, "BEHAVIOR", 8))
@@ -366,7 +366,7 @@ void P_LoadNodes_ZDBSP (int lump, boolean compressed)
     numsubsectors = numSubs;
     subsectors = Z_Malloc(numsubsectors * sizeof(subsector_t), PU_LEVEL, 0);
 
-    for (i = currSeg = 0; i < numsubsectors; i++)
+    for (i = currSeg = 0; i < (unsigned int)numsubsectors; i++)
     {
 	mapsubsector_zdbsp_t *mseg = (mapsubsector_zdbsp_t*) data + i;
 
@@ -391,7 +391,7 @@ void P_LoadNodes_ZDBSP (int lump, boolean compressed)
     numsegs = numSegs;
     segs = Z_Malloc(numsegs * sizeof(seg_t), PU_LEVEL, 0);
 
-    for (i = 0; i < numsegs; i++)
+    for (i = 0; i < (unsigned int)numsegs; i++)
     {
 	line_t *ldef;
 	unsigned int linedef;
@@ -460,7 +460,7 @@ void P_LoadNodes_ZDBSP (int lump, boolean compressed)
     numnodes = numNodes;
     nodes = Z_Malloc(numnodes * sizeof(node_t), PU_LEVEL, 0);
 
-    for (i = 0; i < numnodes; i++)
+    for (i = 0; i < (unsigned int)numnodes; i++)
     {
 	int j, k;
 	node_t *no = nodes + i;
